--- code/src/components/SummaryNameQuestion.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-
-
-export const SummaryNameQuestion = ({ name }) => {
-    return(
-        <div className="summary-container">
-            <div></div>
-            <p>
-            dgdgf
-            {name} 
-            </p>
-        </div> 
-    )
-    }
